move common funcs to utils
